lockers/grace_lock: a better version of grace_lock_checker
The original grace_lock_checker was based on a false semantical
premise: it used a run checker.  The new grace_lock script uses
a staleness checker instead.  Although these merely seem like
opposites, that is not quite accurate.  There are potentially 3
answers to either of those questions: "yes", "no", and "unknown".

It is important to not recover a lock in the "unknown" state.
If one is to ask the question "should I recover", one is asking
the question "is_stale", not "is_running".  Thus if "unknown", one
can answer the question "is_stale" with a "no" without contorting
the truth much:  simply change the question to "is_known_stale".
This makes it easy to write a staleness checker!

However, answering the question "is_running" with "yes" when the
answer is "unknown" is a stretch, and saying "is_known_running"
doesn't help in determining whether to recover.  This semantic
quagmire makes writing run_checkers difficult and potentially
confusing to programmers.

For expected recovery behavior it is important that the "unknown"
corner case be easy to answer.  Fixing the lock_grace_checker
to the better semantics while keeping its API clean requires
breaking its current behavior.  Since it is used in and the original
name is not very consistent with the current naming trend in this
package, it is easier to just create a new lock helper: grace_lock.

Change-Id: I25c8a903611a9260d8112dab1dfed79b2cae4b47
Signed-off-by: Martin Fick <mfick@quicinc.com>