buckets: a task_bucket manager
This keeps a current bucket and hands it out on request.
If the current bucket is closed, a new bucket will be
created and handed out instead.  New buckets can be named
via a pluggable mechanism.  There are also some default
built-in timed based naming schemes such as daily and
hourly.  This makes it very easy to manage a set of tasks
which need to be completed regularly, cooperatively by
multiple actors.

Change-Id: I3bf5afb522976817570f9539823993b7894ffccf
Signed-off-by: Martin Fick <mfick@quicinc.com>