fast_lock: perform wait for cleaning markdirs in parallel
For each markdir present a separate 2s wait was being
performed.  Now find all the present markdirs, wait 2s
and then clean them all up.

This means that some slow creators which previously may
have not been interrupted could now get interrupted.  Those
slow creators previously immune to this were not intentionally
immune to this.  All slow creators have always faced the risk
of being interrupted, it just may happen earlier now.  With this
fix, slow creators are given a 2s window to perform a complete
mkdir -p.  This operation can be slow on a cross-site NFS server.
Testing shows that the entire fast_lock between SD and Boulder
can takes around 1s.  If this ends up causing a problem, the
wait period can be increased for all markers simultaneously
instead of the old policy that allows some to be shorter and
some to be longer.

Change-Id: I85c46fe3bc3a5f0c4de5a4eb236a8fcc91c52f78
Signed-off-by: Martin Fick <mfick@quicinc.com>