fast_lock: rename stale_ids to clean_stale_ids
The new name is more descriptive of what the intent
of this method is, and is more in line with providing
a series of clean functions in the future.  Also
modify all the known callers to use the new name.

This leaves the old function name around and deprecates
it for a removal at a later time.  This phased removal
should help prevent runtime errors during deployments.

Change-Id: I97d43598cef86a7719e6383e758a481402dacfac
Signed-off-by: Martin Fick <mfick@quicinc.com>