grace_lock: add --grace-seconds option
This option was supported by an optional trailing positional parameter.
This made it difficult to set in a more global fashion by callers.
Update the known callers to the new option and remove the API docs for
the old trailing argument option.  The old trailing option is still
functional just in case there are still callers.

Change-Id: Ifc5a3268ac3a39c6017082afcced62a195013b87
Signed-off-by: Martin Fick <mfick@quicinc.com>