grace_lock: loop through checks in a serialized manner
Since we want to be able to move away from unconstrained background
tasks, move to a model where each potentially stale actor is checked
serially within ponder_clean.  This will potentially increase the
runtime of ponder_clean when there are many actors to check for
staleness.  However serialization may actually reduce the overall
workload on the system when contention is high since it reduces the
likelyhood of multiple actors all checking for the same thing when only
one check would do.

Change-Id: Idada39862c08849fee41d9714db02f008666aeb3
Signed-off-by: Martin Fick <mfick@quicinc.com>