grace_lock: constrain the cleanup to 2 checks
By no longer attempting to clean up every potential stale actor, the
system moves to a model where the cleanup cost of a messy system is
spread over multiple API invocations.  This means that a single API
invocation may no longer lead to a completely clean system.  This
tradeoff should lead to more predictable and constrained cleanup
workloads.  This change paves the way for doing all cleanup in the
foreground where it can be serialized with each API call.

Change-Id: Ie666a121bc52239b4940258f4e142b076e6f2bb9
Signed-off-by: Martin Fick <mfick@quicinc.com>