Use build instead of markers dir in fast_lock
'build' is a better name for what we are doing since we are building
proposed future locks.  This is now a pattern that the Gerrit events
plugin does, so use the better name here also.

This change also alter the cleanup of proposals which were not
explicitly done via "clean_stale_ids".  This should not cause any issues
since all the current known users of this script call clean_stale_ids.
This may prevent some lock proposal failures in highly contended
environments.

Lastly, leave 'markers' cleanup behind even though we now use the
'build' dir, to ensure that currently operating locks get cleaned up.
It probably makes sense to remove this support in about a month or so
from deploying it to all known locations.

Change-Id: I300520f2428d9c69dfe60900b7d4f336de6b6f79
Signed-off-by: Martin Fick <mfick@quicinc.com>