ssh_id.sh: validate remote command completed
The ssh_uid command is used to lookup the current id of the remote
running process. If the lookup is believed to be successful and the
remote UID does not match the supplied UID, then it is assumed to be
stale. Add a check to ensure that the lookup was successful, and don't
trust the lookup if the check fails.

Change-Id: I3ed4ac2340fea439f2e87dfbece6741f128aa544
Signed-off-by: Martin Fick <mfick@quicinc.com>