Split generic functionality from lock_ssh into hostpid_lock
lock_ssh was mostly a grace_lock wrapper which understood how to work
with PIDS and translate them to and from more unique IDs which include a
hostid part. Abstract out the bulk of this framework into a building
block of its own so that other hostid based lockers can be easily built.
This should be useful for creating both AWS EC2 and potentially for
Kubernetes based lockers.

Impact: None (refactor only)
Change-Id: I1b304ebfcae4a318c5cdc5863cb24567ec828617
Signed-off-by: Martin Fick <quic_mfick@quicinc.com>
Signed-off-by: Adithya Chakilam <quic_achakila@quicinc.com>