ssh_id: ignore host keys when checking ids via ssh.
A policy of ignoring host keys here should be an acceptable risk for the
use case of this script. In the worst case scenario abuse of this
script, if the host being logged into is impersonated then this script
could be deceived into reporting a false stale id. The advantage is that
the maintenance burden of the ssh based locks should be drastically
reduced.

Change-Id: I9a1b725037aa96082778bcfc436659a5b09ca71f
Impact: host keys will no longer be checked with ssh locks
Risk-Scope: git repacking
Risk-Scope: Gerrit hook queues
Risk-Scope: tracking the need for Gerrit to replicate-all
Signed-off-by: Martin Fick <mfick@quicinc.com>