lock_ssh and test: split host into hostid and sshdest
Previously the term host was being used loosely and sometimes was
referring to the unique hostid as stored in the lock, and sometimes as
the destination to ssh to in order to verify hostids and lock staleness.
Clarify the differences between hostids and ssh destinations by using
different terminology to distinguish the two.

Impact: unused owner_host command was renamed to owner_hostid
Change-Id: Ibf60ce6201f74e09f4fc0277392da51c3244b3f5
Signed-off-by: Martin Fick <mfick@quicinc.com>