Skip notifier tests on nodes not setup for them
Only run the fast notifier tests on nodes which are setup for them. The
tests will not work on nodes were the hostname and fqdn match since they
rely on this missmatch for faking a failure. Some EC2 nodes seem to not
have the fqdn setup, this helps skip these tests on those nodes.

Impact: sometimes tests will warn instead of failing
Change-Id: I533c1a04e3e75272cfcd42d7aa0da8374e528ccd
Signed-off-by: Martin Fick <mfick@quicinc.com>