Introduce K8s locks
This locker allows processes running on different pods in a kubernetes
cluster to create locks on a shared persistent filesystem. Pods are
assumed to have the ability to exec into other pods running on a cluster
to be able to verify the staleness of processes.

Add tests supported to run on a local minikube cluster.

Change-Id: I578fb2c5e2825c1482cbf368884689fdeb61a747
Signed-off-by: Adithya Chakilam <quic_achakila@quicinc.com>